messages-service/requirements.txt
fastapi
uvicorn
kafka-python
python-consul